pyproject.toml
[tool.poetry]
name = "panda-game"
version = "0.1.0"
description = "A fun platformer game where a panda escapes and rescues animal friends"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.14"
pygame = "^2.5.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 